refactor(ai-service): Step 7 — LLM client modules → app/llm/
The two LLM provider modules move out of services/llm/ into the
dedicated outbound-adapter band. The Step 11 split of local_llm into
llama_client.py + vlm_local.py is deferred (1200 lines, not worth
splitting until the ML restructure happens).

Moves:
  app/services/llm/local_llm_service.py  →  app/llm/local_llm.py
  app/services/llm/llama_health.py       →  app/llm/health.py

Importers updated (11 imports across 7 files):
  • app/main.py, app/api/query.py
  • app/services/full_dataset_ml_service.py (module + lazy-inside-function imports)
  • app/services/llm/telemetry_llm_orchestrator.py (module + lazy imports)
  • ui/segment_tabs/components/{llm_ai_annotation, llm_batch_ai_annotation}.py
  • scripts/train_telemetry_llm.py

Skill prompts and label catalog (also under services/llm/ today)
ship to app/skills/ + app/domain/ in Step 8.

Validated: app.main imports clean, /query/llama-health 200,
LocalTelemetryLLM + check_llama_server both reachable via new paths.
