+++ acla_ai_service/app/main.py
@@ -9,7 +9,7 @@ from dotenv import load_dotenv
 
 from app.core import settings
 from app.integrations.backend.client import backend_service
-from app.services.llm.llama_health import check_llama_server
+from app.llm.health import check_llama_server
 from app.api import (
     health_router,
     racing_session_router,
